# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.createTableRequest:
    properties:
      seats_number:
        example: 3
        type: integer
    type: object
  handler.createTableResponse:
    properties:
      id:
        example: 2
        type: integer
    type: object
...
  handler.deleteTableResponse:
    properties:
      status:
        type: string
    type: object
  handler.errResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: missing required fields
        type: string
      status:
        example: invalid request
        type: string
    type: object
...
  handler.getTableResponse:
    properties:
      id:
        example: 3
        type: integer
      restaurant_id:
        example: 2
        type: integer
      seats_number:
        description: SeatsNumber представляет вместимость столика.
        example: 4
        type: integer
    type: object
...
  handler.listTablesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Table'
        type: array
    type: object
...
  handler.updateTableResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  model.Table:
    properties:
      id:
        example: 3
        type: integer
      restaurant_id:
        example: 2
        type: integer
      seats_number:
        description: SeatsNumber представляет вместимость столика.
        example: 4
        type: integer
    type: object
...
  model.UpdateTableData:
    properties:
      seats_number:
        example: 4
        type: integer
    type: object
host: localhost:8080
info:
  contact: { }
  description: API сервиса бронирования столиков в ресторанах
  license:
    name: The MIT License (MIT)
    url: https://mit-license.org/
  termsOfService: http://swagger.io/terms/
  title: Restaurant Table Booking API
  version: "1.0"
paths:
...
  /restaurants/{restaurant_id}/tables/:
    get:
      consumes:
        - application/json
      parameters:
        - description: ID ресторана
          in: path
          name: restaurant_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.listTablesResponse'
        "400":
          description: Некорректный restaurant_id
          schema:
            $ref: '#/definitions/handler.errResponse'
        "500":
          description: Ошибка на стороне сервера
          schema:
            $ref: '#/definitions/handler.errResponse'
      summary: Получить список столиков в ресторане
      tags:
        - tables
    post:
      consumes:
        - application/json
      parameters:
        - description: ID ресторана
          in: path
          name: restaurant_id
          required: true
          type: string
        - description: Информация о столике
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/handler.createTableRequest'
      produces:
        - application/json
      responses:
        "201":
          description: ok
          schema:
            $ref: '#/definitions/handler.createTableResponse'
        "400":
          description: Некорректные данные столика
          schema:
            $ref: '#/definitions/handler.errResponse'
        "500":
          description: Ошибка на стороне сервера
          schema:
            $ref: '#/definitions/handler.errResponse'
      summary: Создать столик в ресторане
      tags:
        - tables
  /tables/{table_id}/:
    delete:
      consumes:
        - application/json
      description: Если за столиком ещё будут сидеть клиенты (есть брони в будущем/сегодняшнем
        днях в этом ресторане), то его нельзя удалить.
      parameters:
        - description: ID столика
          in: path
          name: table_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.deleteTableResponse'
        "400":
          description: Некорректный данные запроса
          schema:
            $ref: '#/definitions/handler.errResponse'
        "500":
          description: Ошибка на стороне сервера
          schema:
            $ref: '#/definitions/handler.errResponse'
      summary: Удаляет столик по его ID
      tags:
        - tables
    get:
      consumes:
        - application/json
      parameters:
        - description: ID столика
          in: path
          name: table_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getTableResponse'
        "400":
          description: Некорректный ID столика
          schema:
            $ref: '#/definitions/handler.errResponse'
        "500":
          description: Ошибка на стороне сервера
          schema:
            $ref: '#/definitions/handler.errResponse'
      summary: Получить столик по его ID
      tags:
        - tables
    patch:
      consumes:
        - application/json
      description: Обновление происходит именно путём PATCH-запросов, чтобы была возможность
        изменять данные частично.
      parameters:
        - description: ID столика
          in: path
          name: table_id
          required: true
          type: string
        - description: Информация о столике
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/model.UpdateTableData'
      produces:
        - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.updateTableResponse'
        "400":
          description: Некорректный данные запроса
          schema:
            $ref: '#/definitions/handler.errResponse'
        "500":
          description: Ошибка на стороне сервера
          schema:
            $ref: '#/definitions/handler.errResponse'
      summary: Обновить информацию о столике по его ID
      tags:
        - tables
swagger: "2.0"